Merged PR 3958: Adding Build agents
the dev cliennt is called metadent. Moved all bicep files for this client into a folder called metadent under dev environment.  This is so that this matches the scheme that will be used in prd (where all clients will be under separate folders.

**Other changes**
- corrected all the naming in the bicep files so that it matches the standard
- vnet subnets.  Created WEB, SQL and DEV subnets such that web and sql work as expected. WEB and SQL subnets required delegation (and so cannot share the DEV subnet)
- web app connected to vnet
- my sql server connected to vnet (not complete as the dns zone doesn't have the servers ip yet)
- added keyvaults with correct naming. added the keyvault default secrets for a client.  autogenerated passwords using newGuid function.
- added storage account.

Examples of changes made:

This should be the format of a client directory:

![image (2).png](https://dev.azure.com/Projectcodeug/cc78b86a-3ce1-4177-a761-eed6f3ebc659/_apis/git/repositories/4d95675f-3f5d-49e1-a012-b9dc4c2b42ba/pullRequests/3958/attachments/image%20%282%29.png)

note the file called "create-dev-clients.ps1" should be able to deploy all clients in an environment.  Whereas the file "create-client.ps1" should be placed in the directory and deploy a single client only

**Database server with database**

![image (3).png](https://dev.azure.com/Projectcodeug/cc78b86a-3ce1-4177-a761-eed6f3ebc659/_apis/git/repositories/4d95675f-3f5d-49e1-a012-b9dc4c2b42ba/pullRequests/3958/attachments/image%20%283%29.png)

Note the server name. This is important because we want to avoid name clashes with the other subscription.  There are some resource types in Azure that have to be globally unique.  One of them is a mysql server name and another is a keyvault name.

**SQL networking bound to a vnet - subnet SQL**

![image (4).png](https://dev.azure.com/Projectcodeug/cc78b86a-3ce1-4177-a761-eed6f3ebc659/_apis/git/repositories/4d95675f-3f5d-49e1-a012-b9dc4c2b42ba/pullRequests/3958/attachments/image%20%284%29.png)

**Metadent client resources**

![image (5).png](https://dev.azure.com/Projectcodeug/cc78b86a-3ce1-4177-a761-eed6f3ebc659/_apis/git/repositories/4d95675f-3f5d-49e1-a012-b9dc4c2b42ba/pullRequests/3958/attachments/image%20%285%29.png)

**Backend app service bound to vnet - subnet WEB**
![image (6).png](https://dev.azure.com/Projectcodeug/cc78b86a-3ce1-4177-a761-eed6f3ebc659/_apis/git/repositories/4d95675f-3f5d-49e1-a012-b9dc4c2b42ba/pullRequests/3958/attachments/image%20%286%29.png)

my PR description has been truncated. I'll post the rest on the channel.